do not inject a JSON::PP dependency as it can never help -- just warn
The very same toolchain configuration that would get as far as reading the
JSON::PP dependency is the one that has just proven that it doesn't need it --
as the dependency is reported in META.json, which the tools just parsed
successfully!  Therefore it is entirely pointless, even for future-proofing.
The best we can do is warn when building a distribution which *might* have
issues.