do not omit empty (file-less) commits from consideration
These are often used to give attribution to people who had patches
submitted with incorrect attribution applied to the commits, or to give thanks
to people who provided help in other ways.

This was regressed in commit 4bba482f, which started always passing a path(s)
to the 'git shortlog' command -- with no paths provided, all commits from the
repository are considered.